2942: Фикс переполнения результатов вычисления

+++ Src/RegionProcess/Download/fr_TilesDownload.pas
@@ -184,7 +184,7 @@ begin
         APixelRect := RectFromDoubleRect(VBounds, rrOutside);
         ATileRect := VProjection.PixelRect2TileRect(APixelRect);
         ATilesCount :=
-          (ATileRect.Right - ATileRect.Left) * (ATileRect.Bottom - ATileRect.Top);
+          Int64(ATileRect.Right - ATileRect.Left) * Int64(ATileRect.Bottom - ATileRect.Top);
         Result := True;
       end;
     end;
